requirements-teams.txt
typer[all]>=0.12.0
rich>=13.7.0
httpx>=0.27.0
anthropic>=0.34.0
pyyaml>=6.0
python-dotenv>=1.0.0
